+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+* added support for artifacts
+* added support for disposition codes 
 * fixed updating user's extension
 * receiving incoming emails & broadcasting attachments, html body, plain text body
 * fixed wrap-up fns